рефакторинг

+++ src/main/java/ru/github/calculator/dto/DepositValueForCalculation.java
@@ -1,4 +1,4 @@
 package ru.github.calculator.dto;
 
-public record DepositValueForCalculation(int sum, int term, int percent) {
+public record DepositValueForCalculation(long sum, long term, long percent) {
 }
